WindowsDomainAuthApp/backend/requirements.txt
flask==2.3.2
flask-cors==4.0.0
python-ldap3==2.9.1
pywin32==306
pyjwt==2.7.0
requests==2.31.0
cryptography==41.0.3
python-dotenv==1.0.0